MAESTRO: address chart range review feedback
